pull code out of main.cpp into application fil
