+++ code/src/hillshade/cpp/hillshade/main.cpp
@@ -1,264 +1,11 @@
 #include <memory>
-#include <iomanip>
-#include <iostream>
 
 #include <Windows.h>
 #include <crtdbg.h>
 
-#include <Common/interface/RefCntAutoPtr.hpp>
-#include <Graphics/GraphicsEngineOpenGL/interface/EngineFactoryOpenGL.h>
-#include <Graphics/GraphicsEngine/interface/RenderDevice.h>
-#include <Graphics/GraphicsEngine/interface/DeviceContext.h>
-#include <Graphics/GraphicsEngine/interface/SwapChain.h>
+#include "hillshade/application.h"
 
-static const char* VSSource = R"(
-struct PSInput 
-{ 
-    float4 Pos   : SV_POSITION; 
-    float3 Color : COLOR; 
-};
-
-void main(in  uint    VertId : SV_VertexID,
-          out PSInput PSIn) 
-{
-    float4 Pos[6];
-    Pos[0] = float4(-0.5, -0.5, 0.0, 1.0);
-    Pos[1] = float4(-0.5, +0.5, 0.0, 1.0);
-    Pos[2] = float4(+0.5, +0.5, 0.0, 1.0);
-    Pos[3] = float4(+0.5, +0.5, 0.0, 1.0);
-    Pos[4] = float4(+0.5, -0.5, 0.0, 1.0);
-    Pos[5] = float4(-0.5, -0.5, 0.0, 1.0);
-
-    float3 Col[6];
-    Col[0] = float3(1.0, 0.0, 0.0);
-    Col[1] = float3(1.0, 1.0, 0.0);
-    Col[2] = float3(0.0, 1.0, 0.0);
-    Col[3] = float3(0.0, 1.0, 0.0);
-    Col[4] = float3(1.0, 1.0, 0.0);
-    Col[5] = float3(1.0, 0.0, 0.0);
-
-    PSIn.Pos   = Pos[VertId];
-    PSIn.Color = Col[VertId];
-}
-)";
-
-// Pixel shader simply outputs interpolated vertex color
-static const char* PSSource = R"(
-struct PSInput 
-{ 
-    float4 Pos   : SV_POSITION; 
-    float3 Color : COLOR; 
-};
-
-struct PSOutput
-{ 
-    float4 Color : SV_TARGET; 
-};
-
-void main(in  PSInput  PSIn,
-          out PSOutput PSOut)
-{
-    PSOut.Color = float4(PSIn.Color.rgb, 1.0);
-}
-)";
-
-
-class Tutorial00App
-{
-public:
-    Tutorial00App()
-    {
-    }
-
-    ~Tutorial00App()
-    {
-        m_pImmediateContext->Flush();
-    }
-
-    bool InitializeDiligentEngine(HWND hWnd)
-    {
-        Diligent::SwapChainDesc SCDesc;
-#    if EXPLICITLY_LOAD_ENGINE_GL_DLL
-        // Load the dll and import GetEngineFactoryOpenGL() function
-        auto GetEngineFactoryOpenGL = Diligent::LoadGraphicsEngineOpenGL();
-#    endif
-        auto* pFactoryOpenGL = GetEngineFactoryOpenGL();
-
-        Diligent::EngineGLCreateInfo EngineCI;
-        EngineCI.Window.hWnd = hWnd;
-
-        pFactoryOpenGL->CreateDeviceAndSwapChainGL(EngineCI, &m_pDevice, &m_pImmediateContext, SCDesc, &m_pSwapChain);
-        return true;
-    }
-
-    void CreateResources()
-    {
-        // Pipeline state object encompasses configuration of all GPU stages
-
-        Diligent::GraphicsPipelineStateCreateInfo PSOCreateInfo;
-
-        // Pipeline state name is used by the engine to report issues.
-        // It is always a good idea to give objects descriptive names.
-        PSOCreateInfo.PSODesc.Name = "Simple triangle PSO";
-
-        // This is a graphics pipeline
-        PSOCreateInfo.PSODesc.PipelineType = Diligent::PIPELINE_TYPE_GRAPHICS;
-
-        // clang-format off
-        // This tutorial will render to a single render target
-        PSOCreateInfo.GraphicsPipeline.NumRenderTargets = 1;
-        // Set render target format which is the format of the swap chain's color buffer
-        PSOCreateInfo.GraphicsPipeline.RTVFormats[0] = m_pSwapChain->GetDesc().ColorBufferFormat;
-        // Use the depth buffer format from the swap chain
-        PSOCreateInfo.GraphicsPipeline.DSVFormat = m_pSwapChain->GetDesc().DepthBufferFormat;
-        // Primitive topology defines what kind of primitives will be rendered by this pipeline state
-        PSOCreateInfo.GraphicsPipeline.PrimitiveTopology = Diligent::PRIMITIVE_TOPOLOGY_TRIANGLE_LIST;
-        // No back face culling for this tutorial
-        PSOCreateInfo.GraphicsPipeline.RasterizerDesc.CullMode = Diligent::CULL_MODE_NONE;
-        // Disable depth testing
-        PSOCreateInfo.GraphicsPipeline.DepthStencilDesc.DepthEnable = Diligent::False;
-        // clang-format on
-
-        Diligent::ShaderCreateInfo ShaderCI;
-        // Tell the system that the shader source code is in HLSL.
-        // For OpenGL, the engine will convert this into GLSL under the hood
-        ShaderCI.SourceLanguage = Diligent::SHADER_SOURCE_LANGUAGE_HLSL;
-        // OpenGL backend requires emulated combined HLSL texture samplers (g_Texture + g_Texture_sampler combination)
-        ShaderCI.Desc.UseCombinedTextureSamplers = true;
-        // Create a vertex shader
-        Diligent::RefCntAutoPtr<Diligent::IShader> pVS;
-        {
-            ShaderCI.Desc.ShaderType = Diligent::SHADER_TYPE_VERTEX;
-            ShaderCI.EntryPoint = "main";
-            ShaderCI.Desc.Name = "Triangle vertex shader";
-            ShaderCI.Source = VSSource;
-            m_pDevice->CreateShader(ShaderCI, &pVS);
-        }
-
-        // Create a pixel shader
-        Diligent::RefCntAutoPtr<Diligent::IShader> pPS;
-        {
-            ShaderCI.Desc.ShaderType = Diligent::SHADER_TYPE_PIXEL;
-            ShaderCI.EntryPoint = "main";
-            ShaderCI.Desc.Name = "Triangle pixel shader";
-            ShaderCI.Source = PSSource;
-            m_pDevice->CreateShader(ShaderCI, &pPS);
-        }
-
-        // Finally, create the pipeline state
-        PSOCreateInfo.pVS = pVS;
-        PSOCreateInfo.pPS = pPS;
-        m_pDevice->CreateGraphicsPipelineState(PSOCreateInfo, &m_pPSO);
-    }
-
-    void Render()
-    {
-        // Set render targets before issuing any draw command.
-        // Note that Present() unbinds the back buffer if it is set as render target.
-        auto* pRTV = m_pSwapChain->GetCurrentBackBufferRTV();
-        auto* pDSV = m_pSwapChain->GetDepthBufferDSV();
-        m_pImmediateContext->SetRenderTargets(1, &pRTV, pDSV, Diligent::RESOURCE_STATE_TRANSITION_MODE_TRANSITION);
-
-        // Clear the back buffer
-        const float ClearColor[] = { 0.350f, 0.350f, 0.350f, 1.0f };
-        // Let the engine perform required state transitions
-        m_pImmediateContext->ClearRenderTarget(pRTV, ClearColor, Diligent::RESOURCE_STATE_TRANSITION_MODE_TRANSITION);
-        m_pImmediateContext->ClearDepthStencil(pDSV, Diligent::CLEAR_DEPTH_FLAG, 1.f, 0, Diligent::RESOURCE_STATE_TRANSITION_MODE_TRANSITION);
-
-        // Set the pipeline state in the immediate context
-        m_pImmediateContext->SetPipelineState(m_pPSO);
-
-        // Typically we should now call CommitShaderResources(), however shaders in this example don't
-        // use any resources.
-
-        Diligent::DrawAttribs drawAttrs;
-        drawAttrs.NumVertices = 6; // Render 6 vertices
-        m_pImmediateContext->Draw(drawAttrs);
-    }
-
-    void Present()
-    {
-        m_pSwapChain->Present();
-    }
-
-    void WindowResize(Diligent::Uint32 Width, Diligent::Uint32 Height)
-    {
-        if (m_pSwapChain)
-            m_pSwapChain->Resize(Width, Height);
-    }
-
-    Diligent::RENDER_DEVICE_TYPE GetDeviceType() const { return m_DeviceType; }
-
-private:
-    Diligent::RefCntAutoPtr<Diligent::IRenderDevice>  m_pDevice;
-    Diligent::RefCntAutoPtr<Diligent::IDeviceContext> m_pImmediateContext;
-    Diligent::RefCntAutoPtr<Diligent::ISwapChain>     m_pSwapChain;
-    Diligent::RefCntAutoPtr<Diligent::IPipelineState> m_pPSO;
-    Diligent::RENDER_DEVICE_TYPE                      m_DeviceType = Diligent::RENDER_DEVICE_TYPE_GL;
-};
-
-std::unique_ptr<Tutorial00App> g_pTheApp;
-
-LRESULT CALLBACK MessageProc(HWND, UINT, WPARAM, LPARAM);
-// Main
-int WINAPI WinMain(_In_ HINSTANCE     hInstance,
-    _In_opt_ HINSTANCE hPrevInstance,
-    _In_ LPSTR         lpCmdLine,
-    _In_ int           nShowCmd)
-{
-#if defined(_DEBUG) || defined(DEBUG)
-    _CrtSetDbgFlag(_CRTDBG_ALLOC_MEM_DF | _CRTDBG_LEAK_CHECK_DF);
-#endif
-
-    g_pTheApp.reset(new Tutorial00App);
-
-    std::string Title("Tutorial00: Hello Win32 (GL)");
-    // Register our window class
-    WNDCLASSEX wcex = { sizeof(WNDCLASSEX), CS_HREDRAW | CS_VREDRAW, MessageProc,
-                       0L, 0L, hInstance, NULL, NULL, NULL, NULL, "SampleApp", NULL };
-    RegisterClassEx(&wcex);
-
-    // Create a window
-    LONG WindowWidth = 1280;
-    LONG WindowHeight = 1024;
-    RECT rc = { 0, 0, WindowWidth, WindowHeight };
-    AdjustWindowRect(&rc, WS_OVERLAPPEDWINDOW, FALSE);
-    HWND wnd = CreateWindow("SampleApp", Title.c_str(),
-        WS_OVERLAPPEDWINDOW, CW_USEDEFAULT, CW_USEDEFAULT,
-        rc.right - rc.left, rc.bottom - rc.top, NULL, NULL, hInstance, NULL);
-    if (!wnd)
-    {
-        MessageBox(NULL, "Cannot create window", "Error", MB_OK | MB_ICONERROR);
-        return 0;
-    }
-    ShowWindow(wnd, nShowCmd);
-    UpdateWindow(wnd);
-
-    if (!g_pTheApp->InitializeDiligentEngine(wnd))
-        return -1;
-
-    g_pTheApp->CreateResources();
-
-    // Main message loop
-    MSG msg = { 0 };
-    while (WM_QUIT != msg.message)
-    {
-        if (PeekMessage(&msg, NULL, 0, 0, PM_REMOVE))
-        {
-            TranslateMessage(&msg);
-            DispatchMessage(&msg);
-        }
-        else
-        {
-            g_pTheApp->Render();
-            g_pTheApp->Present();
-        }
-    }
-
-    g_pTheApp.reset();
-
-    return (int)msg.wParam;
-}
+static std::unique_ptr<hillshade::application> s_app = nullptr;
 
 // Called every time the NativeNativeAppBase receives a message
 LRESULT CALLBACK MessageProc(HWND wnd, UINT message, WPARAM wParam, LPARAM lParam)
@@ -273,9 +20,9 @@ LRESULT CALLBACK MessageProc(HWND wnd, UINT message, WPARAM wParam, LPARAM lPara
         return 0;
     }
     case WM_SIZE: // Window size has been changed
-        if (g_pTheApp)
+        if (s_app)
         {
-            g_pTheApp->WindowResize(LOWORD(lParam), HIWORD(lParam));
+            s_app->resize(LOWORD(lParam), HIWORD(lParam));
         }
         return 0;
 
@@ -300,4 +47,55 @@ LRESULT CALLBACK MessageProc(HWND wnd, UINT message, WPARAM wParam, LPARAM lPara
     default:
         return DefWindowProc(wnd, message, wParam, lParam);
     }
+}
+
+int WINAPI WinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPrevInstance, _In_ LPSTR lpCmdLine, _In_ int nShowCmd)
+{
+#if defined(_DEBUG) || defined(DEBUG)
+    _CrtSetDbgFlag(_CRTDBG_ALLOC_MEM_DF | _CRTDBG_LEAK_CHECK_DF);
+#endif
+
+    // register window
+    WNDCLASSEX wcex = { sizeof(WNDCLASSEX), CS_HREDRAW | CS_VREDRAW, MessageProc, 0L, 0L, hInstance, NULL, NULL, NULL, NULL, "Hillshade", NULL };
+    RegisterClassEx(&wcex);
+
+    // Create a window
+    LONG WindowWidth = 1280;
+    LONG WindowHeight = 1024;
+    RECT rc = { 0, 0, WindowWidth, WindowHeight };
+    AdjustWindowRect(&rc, WS_OVERLAPPEDWINDOW, FALSE);
+    HWND wnd = CreateWindow("Hillshade", "Hillshade", WS_OVERLAPPEDWINDOW, CW_USEDEFAULT, CW_USEDEFAULT, rc.right - rc.left, rc.bottom - rc.top, NULL, NULL, hInstance, NULL);
+    if (!wnd)
+    {
+        MessageBox(NULL, "Cannot create window", "Error", MB_OK | MB_ICONERROR);
+        return 0;
+    }
+    ShowWindow(wnd, nShowCmd);
+    UpdateWindow(wnd);
+
+    s_app = std::make_unique<hillshade::application>();
+    if (!s_app->initialize(wnd))
+        return -1;
+    
+    s_app->create_resources();
+
+    // Main message loop
+    MSG msg = { 0 };
+    while (WM_QUIT != msg.message)
+    {
+        if (PeekMessage(&msg, NULL, 0, 0, PM_REMOVE))
+        {
+            TranslateMessage(&msg);
+            DispatchMessage(&msg);
+        }
+        else
+        {
+            s_app->render();
+            s_app->present();
+        }
+    }
+
+    s_app.reset();
+
+    return (int)msg.wParam;
 }
